requirements.txt
requests~=2.26.0
beautifulsoup4~=4.9.3
pymongo[snappy,srv,tls]==3.11.3
python-dotenv==0.14.0
boto3==1.17.57